+++ src/main/java/dev/muon/dynamic_difficulty/DynamicDifficulty.java
@@ -1,12 +1,17 @@
 package dev.muon.dynamic_difficulty;
 
 import com.mojang.logging.LogUtils;
+import dev.muon.dynamic_difficulty.api.LevelingAPI;
+import dev.muon.dynamic_difficulty.compat.PuffishSkillsProvider;
 import dev.muon.dynamic_difficulty.config.Config;
 import dev.muon.dynamic_difficulty.attribute.ModAttributes;
+import dev.muon.dynamic_difficulty.leveling.PlayerLevelUpdateHandler;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.InterModEnqueueEvent;
+import net.neoforged.fml.loading.LoadingModList;
 import org.slf4j.Logger;
 
 @Mod(DynamicDifficulty.MODID)
@@ -21,5 +26,18 @@ public class DynamicDifficulty {
   public DynamicDifficulty(ModContainer container, IEventBus bus) {
     ModAttributes.REGISTRY.register(bus);
     Config.register(container);
+    bus.addListener(this::onInterMod);
+
+    PlayerLevelUpdateHandler.registerCallback(PlayerLevelUpdateHandler::handlePlayerLevelUpdate);
+  }
+
+  private void onInterMod(InterModEnqueueEvent event) {
+    if (isModLoaded("puffish_skills")) {
+      LevelingAPI.registerPlayerLevelProvider(new PuffishSkillsProvider());
+    }
+  }
+
+  public static boolean isModLoaded (String modId) {
+    return LoadingModList.get().getModFileById(modId) != null;
   }
 }